requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
websockets>=11.0.3
pydantic>=2.4.2
sqlalchemy>=2.0.23
pyyaml>=6.0.1
python-multipart>=0.0.6